Materialize embedded walls in direct Lake tests
